requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
tqdm
opencv-python
openpyxl
scikit-image
tensorboard
torch
torchvision
torchaudio
pyyaml

